requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.18
django-ckeditor==6.7.2
django-filter==24.3
django-js-asset==2.2.0
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.3
typing_extensions==4.12.2
